pythonapp/requirements.txt
python-dotenv
polygon-api-client
google-generativeai
flask
flask-cors
